+++ json_intl/lib/json_intl_data.dart
@@ -1,0 +1,2 @@
+export 'src/json_intl_data.dart';
+export 'src/json_intl_value.dart';