+++ json_intl_gen/lib/generator.dart
@@ -1,0 +1,8 @@
+// Copyright (c) 2020, David PHAM-VAN <dev.nfet.net@gmail.com>
+// All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'package:json_intl_gen/src/generator_options.dart';
+export 'package:json_intl_gen/src/pubspec.dart';
+export 'package:json_intl_gen/src/source_generator.dart';