+++ json_intl_gen/lib/src/pubspec.dart
@@ -3,10 +3,10 @@
 // ignore_for_file: public_member_api_docs
 
 class Pubspec {
-  static const String name = 'json_intl';
+  static const String name = 'json_intl_gen';
 
   static const String description =
-      'Flutter Internationalization library based on Json files and mustache replacement strings.';
+      'Flutter Internationalization library based on Json files and mustache replacement strings. Code generator library.';
 
-  static const String versionFull = '2.0.0';
+  static const String versionFull = '3.0.0';
 }