+++ json_intl_gen/lib/src/source_generator.dart
@@ -6,8 +6,10 @@
 import 'dart:convert';
 
 import 'package:dart_style/dart_style.dart';
-// ignore: implementation_imports
-import 'package:json_intl/src/json_intl_data.dart';
+import 'package:json_intl/json_intl_data.dart';
+
+import 'generator_options.dart';
+import 'generator_utils.dart';
 
 extension _SymbolString on Symbol {
   String getString() {
@@ -20,27 +22,15 @@ extension _SymbolString on Symbol {
 class Generator {
   /// Create a `Generator`
   const Generator({
-    this.defaultLocale = 'en',
-    this.className = 'IntlKeys',
-    this.format = true,
     required this.intl,
-    this.mangle = false,
+    required this.options,
   });
 
-  /// The default locale
-  final String defaultLocale;
-
-  /// The class name to generate
-  final String className;
+  /// Generator options
+  final GeneratorOptions options;
 
   /// The localization strings
   final Map<String, JsonIntlData> intl;
-
-  /// Format the generated dart file
-  final bool format;
-
-  /// Change keys to a random string
-  final bool mangle;
 
   List<String> get _langs {
     final langs = intl.entries.map<String>((e) => e.key).toSet().toList();
@@ -65,30 +55,32 @@ class Generator {
       return _stripExtension(a.substring(i + 1));
     }
 
-    return defaultLocale;
+    return options.defaultLocale;
   }
 
   int _langCompare(String a, String b) {
-    a = _langTag(a)!;
-    b = _langTag(b)!;
-    return a.compareTo(b);
+    final a1 = _langTag(a)!;
+    final b1 = _langTag(b)!;
+    return a1.compareTo(b1);
   }
 
   String _generateName(String key, int len) {
-    if (len > 6) {
-      key = key + len.toString();
-      len = 6;
+    var key_ = key;
+    var len_ = len;
+    if (len_ > 6) {
+      key_ = key_ + len_.toString();
+      len_ = 6;
     }
-    var radix = key.hashCode.toRadixString(36);
+    var radix = key_.hashCode.toRadixString(36);
     if (radix.codeUnitAt(0) < 65) {
       radix = 'n$radix';
     }
-    return radix.substring(0, len);
+    return radix.substring(0, len_);
   }
 
   Iterable<String> _createSourceFromKeys([Map<String, String>? names]) sync* {
     yield '/// Internationalization constants';
-    yield 'class $className {';
+    yield 'mixin ${options.className} {';
 
     final keys = <Symbol>{};
     for (final entry in intl.entries) {
@@ -102,7 +94,7 @@ class Generator {
     final variables = <String>{};
     for (final sortedKey in sortedKeys) {
       final key = sortedKey.getString();
-      var variable = _outputVar(key);
+      var variable = outputVar(key, camelCase: true);
 
       var index = 0;
       final tempVariable = variable;
@@ -116,7 +108,7 @@ class Generator {
       }
 
       var finalName = key;
-      if (mangle) {
+      if (options.mangle) {
         var n = 2;
         do {
           finalName = _generateName(key, n);
@@ -128,7 +120,7 @@ class Generator {
       for (final lang in _langs) {
         final entry = intl[lang]!;
         if (entry.keys.contains(sortedKey)) {
-          yield '  /// ${_langTag(lang)}: ${_outputStr(entry.translate(sortedKey))}';
+          yield '  /// ${_langTag(lang)}: ${outputStr(entry.translate(sortedKey))}';
         } else {
           yield '  /// ${_langTag(lang)}: *** NOT TRANSLATED ***';
         }
@@ -141,37 +133,55 @@ class Generator {
     yield '';
 
     yield '/// Default Locale';
-    yield 'const defaultLocale$className = ${_outputStr(defaultLocale)};';
+    yield 'const defaultLocale${options.className} = ${outputStr(options.defaultLocale)};';
     yield '';
 
     yield '/// Available Locales';
-    yield 'const availableLocales$className = [';
+    yield 'const availableLocales${options.className} = [';
     for (final lang in _langs) {
-      yield '  ${_outputStr(_langTag(lang)!)},';
+      yield '  ${outputStr(_langTag(lang)!)},';
     }
     yield '];';
     yield '';
 
     yield '/// Supported Locales';
-    yield 'const supportedLocales$className = [';
+    yield 'const supportedLocales${options.className} = [';
     for (final lang in _langs) {
-      yield '  Locale(${_outputStr(_langTag(lang)!)}),';
+      yield '  Locale(${outputStr(_langTag(lang)!)}),';
     }
     yield '];';
   }
 
+  String createSource() {
+    return options.builtin
+        ? _createBuiltinFromKeys()
+        : _createSourceFromKeys1();
+  }
+
   /// Generate a dart class from a list of translation keys
-  String createSourceFromKeys() {
+  String _createSourceFromKeys1() {
     final output = <String>[];
 
     output.add('// This file is generated automatically, do not modify');
     output.add('');
     output.add('import \'dart:ui\';');
     output.add('');
+    output.add('import \'package:json_intl/json_intl.dart\';');
+    output.add('');
+
+    output.add('const jsonIntlDelegate = JsonIntlDelegate(');
+    if (options.source != GeneratorOptions.def.source) {
+      output.add('base: ${outputStr(options.source)},');
+    }
+    output.add('availableLocales: availableLocalesIntlKeys,');
+    if (options.debug) {
+      output.add('debug: true,');
+    }
+    output.add(');');
 
     output.addAll(_createSourceFromKeys());
 
-    if (format) {
+    if (options.format) {
       return DartFormatter().format(output.join('\n')).toString();
     }
 
@@ -179,37 +189,44 @@ class Generator {
   }
 
   /// Generate a dart class from a list of translation keys
-  String createBuiltinFromKeys() {
+  String _createBuiltinFromKeys() {
     final output = <String>[];
 
     output.add('// This file is generated automatically, do not modify');
     output.add('');
-    output.add('// ignore_for_file: implementation_imports');
-    output.add('');
     output.add('import \'dart:ui\';');
     output.add('');
-    output.add('import \'package:json_intl/src/json_intl_value.dart\';');
+    output.add('import \'package:json_intl/json_intl.dart\';');
+    output.add('import \'package:json_intl/json_intl_data.dart\';');
     output.add('');
+
+    output.add('const jsonIntlDelegate = JsonIntlDelegateBuiltin(');
+    output.add('data: data${options.className},');
+    output.add('defaultLocale: defaultLocale${options.className},');
+    if (options.debug) {
+      output.add('debug: true,');
+    }
+    output.add(');');
 
     final names = <String, String>{};
     output.addAll(_createSourceFromKeys(names));
 
     output.add('/// Data converted from json strings');
-    output.add('const data$className = {');
+    output.add('const data${options.className} = {');
 
     for (final lang in _langs) {
-      output.add('  ${_outputStr(_langTag(lang)!)}: {');
+      output.add('  ${outputStr(_langTag(lang)!)}: {');
 
       final entry = intl[lang]!;
       final Map<String, dynamic> data = json.decode(entry.toString());
 
       for (final key in data.entries) {
-        output.add('$className.${names[key.key]}: JsonIntlValue({');
+        output.add('${options.className}.${names[key.key]}: JsonIntlValue({');
         for (final gender in key.value.entries) {
           output.add('JsonIntlGender.${gender.key}: {');
           for (final plural in gender.value.entries) {
             output.add(
-                'JsonIntlPlural.${plural.key}: ${_outputStr(plural.value)},');
+                'JsonIntlPlural.${plural.key}: ${outputStr(plural.value)},');
           }
           output.add('},');
         }
@@ -221,38 +238,10 @@ class Generator {
 
     output.add('};');
 
-    if (format) {
+    if (options.format) {
       return DartFormatter().format(output.join('\n')).toString();
     }
 
     return output.join('\n');
-  }
-
-  String _outputVar(String s) {
-    s = s.replaceAll(RegExp(r'[^A-Za-z0-9]'), ' ');
-
-    final group = s.split(RegExp(r'\s+'));
-    final buffer = StringBuffer();
-
-    var first = true;
-    for (var word in group) {
-      if (first) {
-        first = false;
-        buffer.write(word.toLowerCase());
-      } else {
-        buffer.write(word.substring(0, 1).toUpperCase());
-        buffer.write(word.substring(1).toLowerCase());
-      }
-    }
-
-    return buffer.toString();
-  }
-
-  String _outputStr(String s) {
-    s = s.replaceAll(r'\', r'\\');
-    s = s.replaceAll('\n', r'\n');
-    s = s.replaceAll('\r', '');
-    s = s.replaceAll("'", r"\'");
-    return "'$s'";
   }
 }